// client/src/Types/typesForAdminCr.ts
export interface ICrBranProps {
    show: boolean;
    onHide?: any;
}

export interface ICrTypProps {
    show: boolean;
    onHide?: any;
}

export interface ICrDevProps {
    show: boolean;
    onHide?: any;
}

export interface IInfos {
    title?: string;
    description?: string;
    number?: number;
}

export interface IAdminState {
    brandVisible?: boolean;
    typeVisivle?: boolean;
    deviceVisible?: boolean;
}